+++ CarChooser.Domain/PreviousRejection.cs
@@ -2,7 +2,7 @@ namespace CarChooser.Domain
 {
     public class PreviousRejection
     {
-        public long CarId { get; set; }
+        public int CarId { get; set; }
         public RejectionReasons Reason { get; set; }
     }
 }
