+++ CarChooser/Models/RejectionVM.cs
@@ -2,7 +2,7 @@
 {
     public class RejectionVM
     {
-        public long CarId { get; set; }
+        public int CarId { get; set; }
         public string Reason { get; set; }
     }
 }
